src/requirements.txt
python-dotenv==1.0.0
caldav==1.3.2
icalendar==5.0.7
requests==2.31.0
APScheduler==3.10.1
vobject==0.9.6.1
sqlalchemy==2.0.39
loguru==0.7.0
pytz==2023.3
python-dateutil==2.8.2